refactored Product to OrderItem
- logic from OrderProduct constructor was moved to OrderItem
- tests still create OrderProduct because OrderItem is still abstract
